chore: move provider icons under their own module
