+++ apps/ledger-live-mobile/src/icons/providers/ChorusOne.tsx
@@ -42,5 +42,3 @@ export function ChorusOne({ size = 32, outline = "black", ...props }: Props) {
     </Svg>
   );
 }
-
-export default ChorusOne;
