+++ apps/ledger-live-mobile/src/icons/providers/EigenLayer.tsx
@@ -18,5 +18,3 @@ export function EigenLayer({ size = 32, outline = "black", ...props }: Props) {
     </Svg>
   );
 }
-
-export default EigenLayer;
