+++ apps/ledger-live-mobile/src/icons/providers/Lido.tsx
